Fixes review comments.
- Added test description.
- Removed duplicate class.
