--- test-cypress/steps/explore/graphs-visualizations-steps.js
@@ -1,11 +1,0 @@
-const VIEW_URL = '/graphs-visualizations';
-
-export class GraphsVisualizationsSteps {
-    static visit() {
-        cy.visit(VIEW_URL);
-    }
-
-    static verifyUrl() {
-        cy.url().should('include', `${Cypress.config('baseUrl')}${VIEW_URL}`);
-    }
-}
